+++ src/modules/storage/index.ts
@@ -34,7 +34,6 @@ export type {
   RemoveRequest,
   RemoveResponse,
   RemoveResult,
-  RemoveAllRequest,
   RemoveAllResponse,
   RemoveAllResult,
 } from './types';